Refactor test infractructure
